+++ src/primihub/pybind_warpper/express_wrapper.h
@@ -13,7 +13,7 @@ namespace primihub {
 // and will be used by python interpreter.
 class PyMPCExpressExecutor : public MPCExpressExecutor {
 public:
-  PyMPCExpressExecutor(uint8_t party_id);
+  PyMPCExpressExecutor(uint32_t party_id);
   
   // These methods are called by python interpreter, 
   // ensure call them with below order.
@@ -29,14 +29,14 @@ public:
   void importI64ColumnValues(std::string &col_name, py::list &val_list);
   
   // 4. Evaluate express with MPC protocol.
-  void runMPCEvaluate(uint8_t party_id, const std::string &ip,
+  void runMPCEvaluate(uint32_t party_id, const std::string &ip,
                       uint16_t next_port, uint16_t prev_port);
   
   // 5. Reveal MPC result to parties. 
   py::object revealMPCResult(py::list &party_list);
 
 private:
-  uint8_t party_id_;
+  uint32_t party_id_;
 };
 } // namespace primihub
